requirements.txt
pillow
numpy
scipy
torch
tensorboard
safetensors
